Valid HTTP + HTML for exceptions, show traceback
The response to the client in case of an application exception is now
valid HTML and the response includes the Content-Length header to comply
with the specification.

Additionally the traceback is shown if the environment variable DEBUG
is either set to 1 or TRUE (case insensitive) otherwise a general
message is shown (as before).